Add players sitting out (#107)
Order games and players within games correctly
Do not allow removing players from event if they have played a game
Order players in settings correctly.
Fix svelte warnings by using derived
Improve rendering of player names in game card.
Change header and bottom nav to fixed to hopefully solve iOS problem where they move around when overscrolling.
Fixed bug where if you clicked new round, then click back one round, you end up 2 rounds back.

+++ app/Http/Controllers/EventPlayerController.php
@@ -62,6 +62,16 @@ class EventPlayerController extends Controller
      */
     public function destroy(Event $event, Player $player): RedirectResponse
     {
+        $hasGames = $event->games()->whereHas('players', function (Builder $query) use ($player) {
+            $query->where('players.id', $player->id);
+        })->exists();
+
+        if ($hasGames) {
+            Inertia::flash('toast', ['type' => 'error', 'message' => 'Cannot remove player with games in this event.']);
+
+            return back();
+        }
+
         $event->players()->detach($player);
 
         return back();
